Cleanup of the tests so they don't require anything in /secrets to run
Adding log monitoring to the testing so that an unexpected ERROR: or WARNING: log message causes a test failure.

General cleanup.

+++ libblinkstick/test/container_context.py
@@ -2,6 +2,8 @@ import time
 from contextlib import contextmanager
 
 import docker
+import docker.models
+import docker.models.containers
 
 from bitbucket_listener import MongoDataAccess
 from process_context import find_free_port
@@ -16,15 +18,17 @@ def managed_container( *args, **kwargs ):
     client.ping()
     container = None
     try:
-        container = client.containers.run(*args, detach=True, remove=True, **kwargs)
+        container = client.containers.run(*args, detach=True, **kwargs)
+        assert isinstance(container, docker.models.containers.Container)
         while container.status != "running":
             time.sleep(1)
             container = client.containers.get(container.id)
-        # TODO: wait for the container to start
+            assert isinstance(container, docker.models.containers.Container)
         yield ManagedContainer(container)
     finally:
-        if container is not None:
+        if container is not None and isinstance(container, docker.models.containers.Container):
             container.kill()
+            container.remove()
 
 class ManagedMongo(ManagedContainer):
     def __init__(self, container, mongo_port):
@@ -34,11 +38,10 @@ class ManagedMongo(ManagedContainer):
 @contextmanager
 def managed_mongo():
     mongo_port = find_free_port()
-    with managed_container("mongo", ports={'27017/tcp': mongo_port} ) as container:
-        while True:
-            try:
-                MongoDataAccess(f"mongodb://127.0.0.1:{mongo_port}/bitbucket")
-                break
-            except Exception:
-                time.sleep(1)
+    with managed_container("mongo",
+                           ports={"27017": str(mongo_port)}
+                            ) as container:
+        data_access = None
+        while data_access is None or not data_access.has_database():
+            data_access = MongoDataAccess(f"mongodb://127.0.0.1:{mongo_port}/bitbucket")
         yield ManagedMongo(container.container, mongo_port)
